moving callMetadata util to it's own file
